refactor, use `createSlice`
